opt(pop): improve speed of `addInd` class
instead of checking if the pop's specie and the individual species are exactly the same it now only check that the specie's name match. This can introduce a bug if tho differents species with the same name are use in a simulation however, this cases will be very unlikely and the performance gain is quite high.